+++ spoon-client/src/main/java/com/squareup/spoon/Spoon.java
@@ -4,9 +4,10 @@ import android.app.Activity;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
+import android.os.Build;
 import android.os.Looper;
-import android.util.DisplayMetrics;
 import android.util.Log;
+import android.view.View;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -18,6 +19,7 @@ import java.util.regex.Pattern;
 import static android.content.Context.MODE_WORLD_READABLE;
 import static android.graphics.Bitmap.CompressFormat.PNG;
 import static android.graphics.Bitmap.Config.ARGB_8888;
+import static android.os.Environment.getExternalStorageDirectory;
 import static com.squareup.spoon.Chmod.chmodPlusR;
 import static com.squareup.spoon.Chmod.chmodPlusRWX;
 
@@ -25,8 +27,10 @@ import static com.squareup.spoon.Chmod.chmodPlusRWX;
 public final class Spoon {
   static final String SPOON_SCREENSHOTS = "spoon-screenshots";
   static final String NAME_SEPARATOR = "_";
-  static final String TEST_CASE_CLASS = "android.test.InstrumentationTestCase";
-  static final String TEST_CASE_METHOD = "runMethod";
+  static final String TEST_CASE_CLASS_JUNIT_3 = "android.test.InstrumentationTestCase";
+  static final String TEST_CASE_METHOD_JUNIT_3 = "runMethod";
+  static final String TEST_CASE_CLASS_JUNIT_4 = "org.junit.runners.model.FrameworkMethod$1";
+  static final String TEST_CASE_METHOD_JUNIT_4 = "runReflectiveCall";
   private static final String EXTENSION = ".png";
   private static final String TAG = "Spoon";
   private static final Object LOCK = new Object();
@@ -43,11 +47,30 @@ public final class Spoon {
    * @return the image file that was created
    */
   public static File screenshot(Activity activity, String tag) {
+    StackTraceElement testClass = findTestClassTraceElement(Thread.currentThread().getStackTrace());
+    String className = testClass.getClassName().replaceAll("[^A-Za-z0-9._-]", "_");
+    String methodName = testClass.getMethodName();
+    return screenshot(activity, tag, className, methodName);
+  }
+
+  /**
+   * Take a screenshot with the specified tag.  This version allows the caller to manually specify
+   * the test class name and method name.  This is necessary when the screenshot is not called in
+   * the traditional manner.
+   *
+   * @param activity Activity with which to capture a screenshot.
+   * @param tag Unique tag to further identify the screenshot. Must match [a-zA-Z0-9_-]+.
+   * @return the image file that was created
+   */
+  public static File screenshot(Activity activity, String tag, String testClassName,
+      String testMethodName) {
     if (!TAG_VALIDATION.matcher(tag).matches()) {
       throw new IllegalArgumentException("Tag must match " + TAG_VALIDATION.pattern() + ".");
     }
     try {
-      File screenshotDirectory = obtainScreenshotDirectory(activity);
+      File screenshotDirectory =
+          obtainScreenshotDirectory(activity.getApplicationContext(), testClassName,
+              testMethodName);
       String screenshotName = System.currentTimeMillis() + NAME_SEPARATOR + tag + EXTENSION;
       File screenshotFile = new File(screenshotDirectory, screenshotName);
       takeScreenshot(screenshotFile, activity);
@@ -59,8 +82,8 @@ public final class Spoon {
   }
 
   private static void takeScreenshot(File file, final Activity activity) throws IOException {
-    DisplayMetrics dm = activity.getResources().getDisplayMetrics();
-    final Bitmap bitmap = Bitmap.createBitmap(dm.widthPixels, dm.heightPixels, ARGB_8888);
+    View view = activity.getWindow().getDecorView();
+    final Bitmap bitmap = Bitmap.createBitmap(view.getWidth(), view.getHeight(), ARGB_8888);
 
     if (Looper.myLooper() == Looper.getMainLooper()) {
       // On main thread already, Just Do It™.
@@ -105,8 +128,16 @@ public final class Spoon {
     activity.getWindow().getDecorView().draw(canvas);
   }
 
-  private static File obtainScreenshotDirectory(Context context) throws IllegalAccessException {
-    File screenshotsDir = context.getDir(SPOON_SCREENSHOTS, MODE_WORLD_READABLE);
+  private static File obtainScreenshotDirectory(Context context, String testClassName,
+      String testMethodName) throws IllegalAccessException {
+    File screenshotsDir;
+    if (Build.VERSION.SDK_INT >= 21) {
+      // Use external storage.
+      screenshotsDir = new File(getExternalStorageDirectory(), "app_" + SPOON_SCREENSHOTS);
+    } else {
+      // Use internal storage.
+      screenshotsDir = context.getDir(SPOON_SCREENSHOTS, MODE_WORLD_READABLE);
+    }
 
     synchronized (LOCK) {
       if (outputNeedsClear) {
@@ -115,10 +146,8 @@ public final class Spoon {
       }
     }
 
-    StackTraceElement testClass = findTestClassTraceElement(Thread.currentThread().getStackTrace());
-    String className = testClass.getClassName().replaceAll("[^A-Za-z0-9._-]", "_");
-    File dirClass = new File(screenshotsDir, className);
-    File dirMethod = new File(dirClass, testClass.getMethodName());
+    File dirClass = new File(screenshotsDir, testClassName);
+    File dirMethod = new File(dirClass, testMethodName);
     createDir(dirMethod);
     return dirMethod;
   }
@@ -127,8 +156,14 @@ public final class Spoon {
   static StackTraceElement findTestClassTraceElement(StackTraceElement[] trace) {
     for (int i = trace.length - 1; i >= 0; i--) {
       StackTraceElement element = trace[i];
-      if (TEST_CASE_CLASS.equals(element.getClassName()) //
-          && TEST_CASE_METHOD.equals(element.getMethodName())) {
+
+      if (TEST_CASE_CLASS_JUNIT_3.equals(element.getClassName()) //
+          && TEST_CASE_METHOD_JUNIT_3.equals(element.getMethodName())) {
+        return trace[i - 3];
+      }
+
+      if (TEST_CASE_CLASS_JUNIT_4.equals(element.getClassName()) //
+          && TEST_CASE_METHOD_JUNIT_4.equals(element.getMethodName())) {
         return trace[i - 3];
       }
     }