+++ spoon-runner/src/main/java/com/squareup/spoon/SpoonDeviceRunner.java
@@ -2,17 +2,21 @@ package com.squareup.spoon;
 
 import com.android.ddmlib.AdbCommandRejectedException;
 import com.android.ddmlib.AndroidDebugBridge;
+import com.android.ddmlib.CollectingOutputReceiver;
+import com.android.ddmlib.DdmPreferences;
 import com.android.ddmlib.IDevice;
 import com.android.ddmlib.InstallException;
 import com.android.ddmlib.ShellCommandUnresponsiveException;
 import com.android.ddmlib.SyncService;
 import com.android.ddmlib.TimeoutException;
 import com.android.ddmlib.logcat.LogCatMessage;
-import com.android.ddmlib.testrunner.RemoteAndroidTestRunner;
 import com.android.ddmlib.testrunner.IRemoteAndroidTestRunner;
+import com.android.ddmlib.testrunner.ITestRunListener;
+import com.android.ddmlib.testrunner.RemoteAndroidTestRunner;
 import com.google.common.base.Strings;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
+import com.squareup.spoon.adapters.TestIdentifierAdapter;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -26,7 +30,6 @@ import java.util.List;
 import java.util.Map;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.TrueFileFilter;
-import com.squareup.spoon.adapters.TestIdentifierAdapter;
 
 import static com.android.ddmlib.FileListingService.FileEntry;
 import static com.squareup.spoon.Spoon.SPOON_SCREENSHOTS;
@@ -42,6 +45,7 @@ import static com.squareup.spoon.SpoonUtils.obtainRealDevice;
 public final class SpoonDeviceRunner {
   private static final String FILE_EXECUTION = "execution.json";
   private static final String FILE_RESULT = "result.json";
+  private static final String SCREENSHOT_DIR = "app_" + SPOON_SCREENSHOTS;
   static final String TEMP_DIR = "work";
   static final String JUNIT_DIR = "junit-reports";
   static final String IMAGE_DIR = "image";
@@ -62,6 +66,7 @@ public final class SpoonDeviceRunner {
   private final File imageDir;
   private final String classpath;
   private final SpoonInstrumentationInfo instrumentationInfo;
+  private final List<ITestRunListener> testRunListeners;
 
   /**
    * Create a test runner for a single device.
@@ -78,11 +83,12 @@ public final class SpoonDeviceRunner {
    * @param className Test class name to run or {@code null} to run all tests.
    * @param methodName Test method name to run or {@code null} to run all tests.  Must also pass
    *        {@code className}.
+   * @param testRunListeners Additional TestRunListener or empty list.
    */
   SpoonDeviceRunner(File sdk, File apk, File testApk, File output, String serial, boolean debug,
       boolean noAnimations, int adbTimeout, String classpath,
       SpoonInstrumentationInfo instrumentationInfo, String className, String methodName,
-      IRemoteAndroidTestRunner.TestSize testSize) {
+      IRemoteAndroidTestRunner.TestSize testSize, List<ITestRunListener> testRunListeners) {
     this.sdk = sdk;
     this.apk = apk;
     this.testApk = testApk;
@@ -101,6 +107,7 @@ public final class SpoonDeviceRunner {
     this.junitReport = FileUtils.getFile(output, JUNIT_DIR, serial + ".xml");
     this.junitReportDir = new File(output, JUNIT_DIR);
     this.imageDir = FileUtils.getFile(output, IMAGE_DIR, serial);
+    this.testRunListeners = testRunListeners;
   }
 
   /** Serialize to disk and start {@link #main(String...)} in another process. */
@@ -143,7 +150,11 @@ public final class SpoonDeviceRunner {
 
   /** Execute instrumentation on the target device and return a result summary. */
   public DeviceResult run(AndroidDebugBridge adb) {
+
     String appPackage = instrumentationInfo.getApplicationPackage();
+    String testPackage = instrumentationInfo.getInstrumentationPackage();
+    String testRunner = instrumentationInfo.getTestRunnerClass();
+    TestIdentifierAdapter testIdentifierAdapter = TestIdentifierAdapter.fromTestRunner(testRunner);
 
     logDebug(debug, "InstrumentationInfo: [%s]", instrumentationInfo);
 
@@ -160,6 +171,8 @@ public final class SpoonDeviceRunner {
     final DeviceDetails deviceDetails = DeviceDetails.createForDevice(device);
     result.setDeviceDetails(deviceDetails);
     logDebug(debug, "[%s] setDeviceDetails %s", serial, deviceDetails);
+
+    DdmPreferences.setTimeOut(adbTimeout);
 
     try {
       // Now install the main application and the instrumentation application.
@@ -187,34 +200,19 @@ public final class SpoonDeviceRunner {
 
     // Run all the tests! o/
     runTestsOnRunner(result, device);
-logDebug(debug, "done running tests");
+    logDebug(debug, "done running tests");
 
-    // Grab all the parsed logs and map them to individual tests.
-    Map<DeviceTest, List<LogCatMessage>> logs = deviceLogger.getParsedLogs();
-    for (Map.Entry<DeviceTest, List<LogCatMessage>> entry : logs.entrySet()) {
-      DeviceTestResult.Builder builder = result.getMethodResultBuilder(entry.getKey());
-      if (builder != null) {
-        builder.setLog(entry.getValue());
-      }
-    }
+    mapLogsToTests(deviceLogger, result);
 
     try {
       logDebug(debug, "About to grab screenshots and prepare output for [%s]", serial);
+      pullScreenshotsFromDevice(device);
 
-      // Sync device screenshots, if any, to the local filesystem.
-      String dirName = "app_" + SPOON_SCREENSHOTS;
-      String localDirName = work.getAbsolutePath();
-      final String devicePath = "/data/data/" + appPackage + "/" + dirName;
-      FileEntry deviceDir = obtainDirectoryFileEntry(devicePath);
-      logDebug(debug, "Pulling screenshots from [%s] %s", serial, devicePath);
-
-      device.getSyncService()
-          .pull(new FileEntry[] {deviceDir}, localDirName, SyncService.getNullProgressMonitor());
-
-      File screenshotDir = new File(work, dirName);
+      File screenshotDir = new File(work, SCREENSHOT_DIR);
       if (screenshotDir.exists()) {
         imageDir.mkdirs();
 
+        logDebug(debug, "Moving screenshots to the image folder on [%s]", serial);
         // Move all children of the screenshot directory into the image folder.
         File[] classNameDirs = screenshotDir.listFiles();
         if (classNameDirs != null) {
@@ -244,6 +242,7 @@ logDebug(debug, "done running tests");
             }
           }
 
+          logDebug(debug, "Generating animated gifs for [%s]", serial);
           // Don't generate animations if the switch is present
           if (!noAnimations) {
             // Make animated GIFs for all the tests which have screenshots.
@@ -264,6 +263,7 @@ logDebug(debug, "done running tests");
     } catch (Exception e) {
       result.addException(e);
     }
+    logDebug(debug, "Done running for [%s]", serial);
 
     return result.build();
   }
@@ -298,10 +298,8 @@ logDebug(debug, "done running tests");
         runTestInBatches(testPackage, testRunner, device, testIdentifierAdapter, result,
                 testsProvider);
       } else {
-        runner.run(
-          new SpoonTestRunListener(result, debug, testIdentifierAdapter),
-          new XmlTestRunListener(junitReport)
-        );
+        List<ITestRunListener> listeners = getiTestRunListeners(result, testIdentifierAdapter);
+        runner.run(listeners);
       }
     } catch (Exception e) {
       logDebug(debug, "caught exception %s", e);
@@ -309,6 +307,16 @@ logDebug(debug, "done running tests");
       result.addException(e);
     }
   }
+
+    private List<ITestRunListener> getiTestRunListeners(DeviceResult.Builder result, TestIdentifierAdapter testIdentifierAdapter) {
+      List<ITestRunListener> listeners = new ArrayList<ITestRunListener>();
+      listeners.add(new SpoonTestRunListener(result, debug, testIdentifierAdapter));
+      listeners.add(new XmlTestRunListener(junitReport));
+      if (testRunListeners != null) {
+        listeners.addAll(testRunListeners);
+      }
+      return listeners;
+    }
 
     // TODO add  comment why we need to run in batches
   private void runTestInBatches(String testPackage,
@@ -335,16 +343,74 @@ logDebug(debug, "done running tests");
       File f = FileUtils.getFile(junitReportDir, serial
                 + "_" + String.valueOf(run) + ".xml");
       logDebug(debug, "logging JUnit to %s", f.getPath());
-      runner.run(
-            new SpoonTestRunListener(result, debug, testIdentifierAdapter),
-            new XmlTestRunListener(f)
-      );
+      List<ITestRunListener> listeners = getiTestRunListeners(result, testIdentifierAdapter);
+      runner.run(listeners);
+
       cls = testsProvider.getNextTests(testCount);
       ++run;
     }
   }
 
-    /////////////////////////////////////////////////////////////////////////////
+  /////////////////////////////////////////////////////////////////////////////
+  /** Download all screenshots from a single device to the local machine. */
+  private void pullScreenshotsFromDevice(IDevice device) throws Exception {
+    // Screenshot path on private internal storage, for KitKat and below.
+    FileEntry internalDir = getScreenshotDirOnInternalStorage();
+    logDebug(debug, "Internal path is " + internalDir.getFullPath());
+
+    // Screenshot path on public external storage, for Lollipop and above.
+    FileEntry externalDir = getScreenshotDirOnExternalStorage(device);
+    logDebug(debug, "External path is " + externalDir.getFullPath());
+
+    // Sync device screenshots to the local filesystem.
+    // TODO only pull from one location, based on android version of device
+    logDebug(debug, "Pulling screenshots from external dir on [%s]", serial);
+    String localDirName = work.getAbsolutePath();
+    adbPull(device, externalDir, localDirName);
+    logDebug(debug, "Pulling screenshots from internal dir on [%s]", serial);
+    adbPull(device, internalDir, localDirName);
+    logDebug(debug, "Done pulling screenshots from [%s]", serial);
+  }
+
+  private void adbPull(IDevice device, FileEntry remoteDirName, String localDirName) {
+    try {
+      device.getSyncService()
+          .pull(new FileEntry[] {remoteDirName}, localDirName,
+              SyncService.getNullProgressMonitor());
+    } catch (Exception e) {
+      logDebug(debug, e.getMessage(), e);
+    }
+  }
+
+  private FileEntry getScreenshotDirOnInternalStorage() {
+    String appPackage = instrumentationInfo.getApplicationPackage();
+    String internalPath = "/data/data/" + appPackage + "/" + SCREENSHOT_DIR;
+    return obtainDirectoryFileEntry(internalPath);
+  }
+
+  private static FileEntry getScreenshotDirOnExternalStorage(IDevice device) throws Exception {
+    String externalPath = getExternalStoragePath(device) + "/" + SCREENSHOT_DIR;
+    return obtainDirectoryFileEntry(externalPath);
+  }
+
+  private static String getExternalStoragePath(IDevice device) throws Exception {
+    CollectingOutputReceiver pathNameOutputReceiver = new CollectingOutputReceiver();
+    device.executeShellCommand("echo $EXTERNAL_STORAGE", pathNameOutputReceiver);
+    return pathNameOutputReceiver.getOutput().trim();
+  }
+
+  /** Grab all the parsed logs and map them to individual tests. */
+  private static void mapLogsToTests(SpoonDeviceLogger deviceLogger, DeviceResult.Builder result) {
+    Map<DeviceTest, List<LogCatMessage>> logs = deviceLogger.getParsedLogs();
+    for (Map.Entry<DeviceTest, List<LogCatMessage>> entry : logs.entrySet()) {
+      DeviceTestResult.Builder builder = result.getMethodResultBuilder(entry.getKey());
+      if (builder != null) {
+        builder.setLog(entry.getValue());
+      }
+    }
+  }
+
+  /////////////////////////////////////////////////////////////////////////////
   ////  Secondary Per-Device Process  /////////////////////////////////////////
   /////////////////////////////////////////////////////////////////////////////